Init graphic fixes about derein review;
Several changes were made to incorporate the ApplicationUrl into various parts of the codebase. Firstly, it was added to the PortfolioSettingsViewModel for use in creating links. This value was then populated in the PortfolioHelper.

In addition, changes were made to numerous views to adjust styling for better usability and to handle link creation. For instance, in the List.cshtml within the Portfolio views, logic was adjusted to open the portfolio preview in a new tab and provide a copy link feature.

Some adjustments for improved user experience were also made, such as modifying display options in the Edit.cshtml within the Cloud views. Related changes to the CSS files were done to complement these modifications.

Furthermore, tweaks were made in the layout views to finesse the display of user-related information.

Lastly, UserDescription was added to the DashboardViewModel for displaying user's description.